Runde av utledede månedsinntekter etter standard avrundingsregel

+++ src/main/kotlin/no/nav/bidrag/inntekt/util/InntektUtil.kt
@@ -3,6 +3,7 @@ package no.nav.bidrag.inntekt.util
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
+import java.math.BigDecimal
 import java.text.SimpleDateFormat
 import java.time.LocalDate
 import java.time.Month
@@ -66,10 +67,10 @@ fun String.isNumeric(): Boolean {
     return this.all { it.isDigit() }
 }
 
-fun beregneBeløpPerMåned(beløp: Int, antallMnd: Int): Int {
+fun beregneBeløpPerMåned(beløp: Int, antallMnd: Int): BigDecimal {
     return if (antallMnd == 0) {
-        0
+        BigDecimal.valueOf(0)
     } else {
-        beløp.div(antallMnd)
+        beløp.toBigDecimal().div(antallMnd.toBigDecimal())
     }
 }
